+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.44.0] - 2023-07-13
+
 ### Added
 
 - Add a function for calculating the size limit of log volumes ([#621]).